pyproject.toml
[tool.poetry]
name = "mfanalyser"
version = "0.1.0"
description = "MF Analyzer using crewAI"
authors = ["Sudhir Dharan <sudhir.dharan@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = "^0.85.0" }
selenium = "^4.21.0"
google-generativeai = "^0.8.4"

[tool.poetry.scripts]
mfanalyser = "mfanalyser.main:run"
train = "mfanalyser.main:train"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
